Closes #31. Creación del fichero para crear la matriz de costos.
